[todo-backend] Add integration tests for using bootable jar
The ToDoIT integration tests was using the regular Arquillian WildFly
container that tested it on a standalone server. This was not correct as
the todo-backend uses a specific provisioning and configuration to run
the application.

The ToDoIT tests now relieson the Arquilliant Bootable Jar container to
ensures that it is tested with the actual bootable jar artifact.
We use the docker-maven-plugin to start/stop a PostgreSQL database
before running the integration tests.

Signed-off-by: Jeff Mesnil <jmesnil@redhat.com>